Add RegionController implementation and supprimerCommune method in CommuneController
